Continuously add all containers to queue.
The reason for this change is for caching performance. Previously we add
one by one, we wait until the next tick to add another container. This
isn't very efficient. Instead, in this version we add all containers at
the same time. So the warmer can burn through all the containers very
fast. There is no randomisation of the container ordering, so there is
some possibility of repetition, but I can't think of any situations
where this would be a major problem.