Remove image fetch queue
We want more or less these properties of the cache warmer:

 1. refreshes the cache as often as rate limiting allows (does all the
    work it can, given backpressure)
 2. does not spin freely (do no more work than needed)

It turns out that a queue is not necessary for having these
properties. We can more simply fill the cache, passing over entries we
already have that are not near expiry, and ask for more entries when
that's done. To avoid spinning (in this case, continuously asking the
cluster for the images to fetch), we can rate limit upstream requests
using a ticker.

The commit also removes some bookkeeping in the warm method, by doing
the fetch and store into memcache in one go(routine); and, by starting
a goroutine per fetch rather than pre-allocating them.